Change folder names and address comments
