+++ in-vehicle-stack/interfaces/invehicle_digital_twin/v1/invehicle_digital_twin.proto
