+++ in-vehicle-stack/interfaces/module/managed_subscribe/v1/managed_subscribe.proto
