+++ in-vehicle-stack/interfaces/service_discovery/v1/service_registry.proto
