+++ scenarios/smart_trailer_use_case/digital-twin-model/dtdl/trailer.json
