+++ scenarios/smart_trailer_use_case/digital-twin-model/src/lib.rs
