+++ scenarios/smart_trailer_use_case/interfaces/digital_twin_get_provider.proto
@@ -1,0 +1,26 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+// SPDX-License-Identifier: MIT
+
+// Digital Twin "Get" Provider definition
+//
+// The protobuf definitions for a Digital Twin Provider which only supports synchronous
+// "Get" operation
+
+syntax = "proto3";
+package digital_twin_get_provider;
+
+// The service entry point to the Digital Twin Get Provider. This simple provider has one method
+// to get the property
+service DigitalTwinGetProvider {
+  // Method which gets the value of the specified property
+  rpc Get (GetRequest) returns (GetResponse);
+}
+
+message GetRequest {
+  string entity_id = 1;
+}
+
+message GetResponse {
+  bool property_value = 1;
+}
