+++ scenarios/smart_trailer_use_case/proto_build/Cargo.toml
