+++ scenarios/smart_trailer_use_case/proto_build/build.rs
@@ -3,6 +3,6 @@
 // SPDX-License-Identifier: MIT
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::compile_protos("../interfaces/trailer_connected_provider.proto")?;
+    tonic_build::compile_protos("../interfaces/digital_twin_get_provider.proto")?;
     Ok(())
 }
