+++ scenarios/smart_trailer_use_case/proto_build/src/lib.rs
@@ -2,9 +2,9 @@
 // Licensed under the MIT license.
 // SPDX-License-Identifier: MIT
 
-pub mod trailer_connected_provider {
+pub mod digital_twin_get_provider {
     pub mod v1 {
         #![allow(clippy::derive_partial_eq_without_eq)]
-        tonic::include_proto!("trailer_connected_provider");
+        tonic::include_proto!("digital_twin_get_provider");
     }
 }
